Added SdlGui and SdlWidget

+++ src/GUI/SdlGui.h
@@ -1,13 +1,15 @@
 #ifndef _H_GUI_H_
 #define _H_GUI_H_
 
+#include "SdlWidget.h"
+
 class SdlGui
 {
     public:
         SdlGui & getObject();
         void addWidget(SdlWidget & widget);
         void display();
-        void update(Uint16 mmsec);
+        void update(unsigned int mmsec);
         void checkInput();
 };
